requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
Deprecated==1.2.14
Django==5.0.7
django-phone-verify==3.0.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
mysqlclient==2.2.4
nexmo==2.5.2
phonenumbers==8.13.42
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
sqlparse==0.5.1
twilio==9.2.3
tzdata==2024.1
urllib3==2.2.2
wrapt==1.16.0
yarl==1.9.4
